+++ packages/bundles/src/plugin-refresh.ts
@@ -1,0 +1,3 @@
+import RefreshPlugin from '../compiled/@rspack/plugin-react-refresh/dist/index.js';
+
+export default RefreshPlugin;