+++ packages/ice/bin/ice-cli.mjs
@@ -6,8 +6,6 @@ import fse from 'fs-extra';
 import { fileURLToPath } from 'url';
 import { program, Option } from 'commander';
 import yargsParser from 'yargs-parser';
-// hijack webpack before import other modules
-import '../esm/requireHook.js';
 import createService from '../esm/createService.js';
 import { TARGETS, WEB } from '../esm/constant.js';
 
@@ -38,7 +36,11 @@ const __dirname = path.dirname(fileURLToPath(import.meta.url));
         ...parseUnknownOptions(ctx.args),
         ...commandArgs,
       } });
-      service.run();
+      service.run().catch((error) => {
+        console.log(chalk.red('Build Error'), error);
+        // Set exit code to 1 to exit process with failure code, otherwise CI may regard build as success.
+        process.exit(1);
+      });
     });
 
   program