+++ packages/runtime/src/types.ts
@@ -3,7 +3,7 @@ import type { InitialEntry, AgnosticRouteObject, Location, History, RouterInit, 
 import type { ComponentType, PropsWithChildren } from 'react';
 import type { HydrationOptions, Root } from 'react-dom/client';
 import type { Params, RouteObject } from 'react-router-dom';
-import type { RenderToPipeableStreamOptions } from 'react-dom/server';
+import type { RenderToPipeableStreamOptions } from './server/streamRender.js';
 import type { RouteLoaderOptions } from './routes.js';
 
 type UseConfig = () => RouteConfig<Record<string, any>>;
@@ -322,7 +322,8 @@ export interface ServerRenderOptions {
   createRoutes: (options: Pick<RouteLoaderOptions, 'requestContext' | 'renderMode'>) => RouteItem[];
   runtimeModules: RuntimeModules;
   documentDataLoader?: DocumentDataLoaderConfig;
-  Document: DocumentComponent;
+  preRender?: boolean;
+  Document?: DocumentComponent;
   documentOnly?: boolean;
   renderMode?: RenderMode;
   // basename is used both for server and client, once set, it will be sync to client.