+++ packages/shared-config/src/index.ts
@@ -1,16 +1,20 @@
-import compilationPlugin, { isSupportedFeature } from './unPlugins/compilation.js';
-import compileExcludes from './compileExcludes.js';
+import compilationPlugin, { isSupportedFeature, getJsxTransformOptions } from './unPlugins/compilation.js';
+import compileExcludes, { SKIP_COMPILE as skipCompilePackages } from './compileExcludes.js';
 import getCompilerPlugins from './getCompilerPlugins.js';
 import getDefineVars from './getDefineVars.js';
 import getPostcssOpts from './getPostcssOpts.js';
 import getCSSModuleLocalIdent from './getCSSModuleLocalIdent.js';
+import getAliasWithRoot from './getAlias.js';
 
 export {
   getCSSModuleLocalIdent,
   compilationPlugin,
   isSupportedFeature,
   compileExcludes,
+  skipCompilePackages,
+  getJsxTransformOptions,
   getCompilerPlugins,
   getDefineVars,
   getPostcssOpts,
+  getAliasWithRoot,
 };