+++ packages/plugin-rax-compat/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [a68ff48d]
+  - @ice/bundles@0.2.0
+
 ## 0.2.9
 
 ### Patch Changes